git commit

+++ Frontend/front/src/pages/profile.js
@@ -1,11 +1,9 @@
-// profile.js
-
 import React, { useState, useEffect } from 'react';
-import { Row, Col, Form, Button } from 'react-bootstrap';
+import { Row, Col, Form, Button, Modal } from 'react-bootstrap';
 import axios from 'axios';
 import './profile.css';
 
-const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
+const Profile = ({ user, onUpdateProfile, onDeleteProfile }) => {
   const [userData, setUserData] = useState({
     firstName: '',
     lastName: '',
@@ -15,59 +13,75 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
   });
 
   const [isFormEdited, setIsFormEdited] = useState(false);
+  const [isEditMode, setIsEditMode] = useState(false);
+  const [showModal, setShowModal] = useState(false);
 
   useEffect(() => {
     if (user) {
-      axios.get(`http://localhost:3001/api/signup/${user}`)
-        .then(response => {
+      axios
+        .get(`http://localhost:3001/api/signup/${user}`)
+        .then((response) => {
           const userDataFromServer = response.data;
           setUserData(userDataFromServer);
         })
-        .catch(error => {
+        .catch((error) => {
           console.error('Error fetching user data:', error);
         });
     }
   }, [user]);
 
   const handleInputChange = (e) => {
-    const { name, value } = e.target;
-    setUserData((prevData) => ({ ...prevData, [name]: value }));
-    setIsFormEdited(true); // Set form as edited when input changes
+    if (isEditMode) {
+      const { name, value } = e.target;
+      setUserData((prevData) => ({ ...prevData, [name]: value }));
+      setIsFormEdited(true);
+    }
   };
 
   const handleSaveProfile = () => {
-    axios.put(`http://localhost:3001/api/signup/${user}`, userData)
-      .then(response => {
+    axios
+      .put(`http://localhost:3001/api/signup/${user}`, userData)
+      .then((response) => {
         console.log('Profile updated successfully:', response.data);
-        onUpdateProfile(response.data);  // Update with the data received from the server
-        setIsFormEdited(false); // Reset form edited status after saving
+        onUpdateProfile(response.data);
+        setIsFormEdited(false);
+        setIsEditMode(false);
+        setShowModal(true); // Show the modal after profile update
       })
-      .catch(error => {
+      .catch((error) => {
         console.error('Error updating profile:', error);
       });
   };
+  
 
   const handleDeleteProfile = () => {
-    axios.delete(`http://localhost:3001/api/signup/${user}`)
-      .then(response => {
+    axios
+      .delete(`http://localhost:3001/api/signup/${user}`)
+      .then((response) => {
         console.log('Profile deleted successfully:', response.data);
-        onDeleteProfile(); // Notify parent component about deletion
+        onDeleteProfile();
       })
-      .catch(error => {
+      .catch((error) => {
         console.error('Error deleting profile:', error);
       });
   };
+
+  const handleCloseModal = () => {
+  setShowModal(false);
+  // Perform any other cleanup or actions needed after closing the modal
+};
+
 
   return (
     <div className="profile-wrapper" id="profile">
       <Row>
         <Col>
-          <h2 className='profile-title'>Customer Profile</h2>
+          <h2 className="profile-title">User Profile</h2>
         </Col>
       </Row>
       <Row>
         <Col className="mx-auto">
-          <Form>
+          <Form className='profile-form'> 
             <Form.Group as={Row} controlId="formName" className="mb-3">
               <Form.Label column sm={2}>
                 First Name
@@ -79,10 +93,11 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
                   name="firstName"
                   value={userData.firstName}
                   onChange={handleInputChange}
+                  readOnly={!isEditMode}
                 />
               </Col>
             </Form.Group>
-            <Form.Group as={Row} controlId="formlastname" className="mb-3">
+            <Form.Group as={Row} controlId="formLastName" className="mb-3">
               <Form.Label column sm={2}>
                 Last Name
               </Form.Label>
@@ -93,10 +108,11 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
                   name="lastName"
                   value={userData.lastName}
                   onChange={handleInputChange}
+                  readOnly={!isEditMode}
                 />
               </Col>
             </Form.Group>
-            <Form.Group as={Row} controlId="formemail" className="mb-3">
+            <Form.Group as={Row} controlId="formEmail" className="mb-3">
               <Form.Label column sm={2}>
                 Email
               </Form.Label>
@@ -107,10 +123,11 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
                   name="email"
                   value={userData.email}
                   onChange={handleInputChange}
+                  readOnly={!isEditMode}
                 />
               </Col>
             </Form.Group>
-            <Form.Group as={Row} controlId="formusername" className="mb-3">
+            <Form.Group as={Row} controlId="formUsername" className="mb-3">
               <Form.Label column sm={2}>
                 Username
               </Form.Label>
@@ -121,6 +138,7 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
                   name="username"
                   value={userData.username}
                   onChange={handleInputChange}
+                  readOnly={!isEditMode}
                 />
               </Col>
             </Form.Group>
@@ -135,6 +153,7 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
                   name="password"
                   value={userData.password}
                   onChange={handleInputChange}
+                  readOnly={!isEditMode}
                 />
               </Col>
             </Form.Group>
@@ -144,17 +163,34 @@ const Profile = ({ user, onUpdateProfile ,onDeleteProfile}) => {
       <Row>
         <Col className="text-right">
           <div className="d-flex align-items-center">
-            {isFormEdited && (
-              <Button variant="primary" className="mr-3" onClick={handleSaveProfile}>
+            {!isEditMode && (
+              <Button className="mr-4" onClick={() => setIsEditMode(true)}>
+                Edit
+              </Button>
+            )}
+            {isFormEdited && isEditMode && (
+              <Button className="mr-4 " onClick={handleSaveProfile}>
                 Save
               </Button>
             )}
-             <Button className="mr-4" onClick={handleDeleteProfile}>
+            <Button className="mr-4" onClick={handleDeleteProfile}>
               Delete Profile
             </Button>
           </div>
         </Col>
       </Row>
+      <Modal show={showModal} onHide={handleCloseModal} centered>
+        <Modal.Header closeButton>
+          <Modal.Title>Profile Updated</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>Your profile has been successfully updated.</Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={handleCloseModal}>
+            Close
+          </Button>
+        </Modal.Footer>
+      </Modal>
+    
     </div>
   );
 };
